backend/roboflow/requirements.txt
ultralytics==8.3.0
gradio==4.44.0
torch==2.5.0
torchvision==0.20.0
opencv-python-headless==4.10.0.84
numpy==1.26.4
Pillow==10.4.0
